process node with html escape
